+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 - Motion fixes in `DetailEntryBottomSheet`
 - Search
   - Search Data Layer
+  - Search UI Layer
+- Link Icon